Add PON ME Explorer page to base variant WebUI
- Replace pontop_order(), pontop_labels() and pontop_mods() with pontop_page_details()
- Change mib_file configuration to select box
- Add 8311_hostname fwenv to set the system hostname
- Implement the 8311_console_en and 8311_dying_gasp_en fwenvs
- Add telus mib and cleaner 1U mib
- Rename 40-8311 uci-defaults file to zzz-8311 so that it runs last
- Add fwenv_get and fwenv_set helper scripts for getting/setting fwenvs. fwenv_set only needs to be run once
- Remove a number of less-than-useful pages from the basic variant WebUI

+++ files/common/lib/8311.sh
@@ -21,21 +21,29 @@ strtolower() {
 	tr '[A-Z]' '[a-z]'
 }
 
+
+fwenv_get_8311() {
+	fwenv_get "8311_$1"
+}
+
+fwenv_set_8311() {
+	fwenv_set "8311_$1" "$2"
+}
+
 _8311_check_persistent_root() {
 	# Remove persistent root
-	PERSIST_ROOT=$(fw_printenv -n 8311_persist_root 2>/dev/null)
+	PERSIST_ROOT=$(fwenv_get_8311 "persist_root")
 	if ! { [ "$PERSIST_ROOT" -eq "1" ] 2>/dev/null; }; then
 		echo "8311_persist_root not enabled, checking bootcmd..." | to_console
-		BOOTCMD=$(fw_printenv -n bootcmd 2>/dev/null)
+		BOOTCMD=$(fwenv_get "bootcmd")
 		if ! { echo "$BOOTCMD" | grep -Eq '^\s*run\s+ubi_init\s*;\s*ubi\s+remove\s+rootfs_data\s*;\s*run\s+flash_flash\s*$'; }; then
 			echo "Resetting bootcmd to default value and rebooting, set fwenv 8311_persist_root=1 to avoid this" | tee -a /dev/console
 
-			fw_setenv bootcmd "run ubi_init; ubi remove rootfs_data; run flash_flash"
-			fw_setenv bootcmd "run ubi_init; ubi remove rootfs_data; run flash_flash"
+			fwenv_set bootcmd "run ubi_init; ubi remove rootfs_data; run flash_flash"
 
 			sync
 			reboot
-			sleep 15
+			sleep 60
 		fi
 	fi
 }
@@ -57,8 +65,16 @@ _mib_update() {
 	sed -r "s#^(([!?]\s+)?${ME}(\s+$P){${PARAM}}\s+)($P)#\1${VALUE}#g" -i "$(_mib_file)"
 }
 
+get_8311_console_en() {
+	fwenv_get_8311 "console_en"
+}
+
+get_8311_dying_gasp_en() {
+	fwenv_get_8311 "dying_gasp_en"
+}
+
 get_8311_gpon_sn() {
-	fw_printenv -n "8311_gpon_sn" 2>/dev/null | grep -E '^[A-Za-z0-9]{4}[A-Fa-f0-9]{8}$'
+	fwenv_get_8311 "gpon_sn" | grep -E '^[A-Za-z0-9]{4}[A-Fa-f0-9]{8}$'
 }
 
 set_8311_gpon_sn() {
@@ -67,7 +83,7 @@ set_8311_gpon_sn() {
 }
 
 get_8311_device_sn() {
-	fw_printenv -n "8311_device_sn" 2>/dev/null
+	fwenv_get_8311 "device_sn" 2>/dev/null
 }
 
 set_8311_device_sn() {
@@ -76,7 +92,7 @@ set_8311_device_sn() {
 }
 
 get_8311_vendor_id() {
-	fw_printenv -n "8311_vendor_id" 2>/dev/null | grep -E '^[A-Za-z0-9]{4}$'
+	fwenv_get_8311 "vendor_id" | grep -E '^[A-Za-z0-9]{4}$'
 }
 
 set_8311_vendor_id() {
@@ -88,7 +104,7 @@ set_8311_vendor_id() {
 }
 
 get_8311_mib_file() {
-	local tmp=$(fw_printenv -n 8311_mib_file 2>/dev/null)
+	local tmp=$(fwenv_get_8311 "mib_file")
 	local mib='/etc/mibs/prx300_1U.ini'
 	if [ -n "$tmp" ]; then
 		if [ -f "/etc/mibs/$tmp" ]; then
@@ -111,7 +127,7 @@ set_8311_mib_file() {
 }
 
 get_8311_reg_id_hex() {
-	{ { fw_printenv -n "8311_reg_id_hex" | hex2str; } || echo -n "$(fw_printenv -n 8311_reg_id)"; cat /dev/zero; } 2>/dev/null | head -c 36 | str2hex | strtoupper
+	{ { fwenv_get_8311 "reg_id_hex" | hex2str; } || echo -n "$(fwenv_get_8311 "reg_id")"; cat /dev/zero; } 2>/dev/null | head -c 36 | str2hex | strtoupper
 }
 
 
@@ -128,7 +144,7 @@ get_8311_sw_ver() {
 	{ [ "$1" = "A" ] || [ "$1" = "B" ]; } && i1="$1"
 	[ "$image1" = "B" ] && i2="A"
 	
-	{ fw_printenv -n "8311_sw_ver$i1" || fw_printenv -n "8311_sw_ver" || fw_printenv -n "8311_sw_ver$i2"; } 2>/dev/null | head -c 14
+	{ fwenv_get_8311 "sw_ver$i1" || fwenv_get_8311 "sw_ver" || fwenv_get_8311 "sw_ver$i2"; } | head -c 14
 }
 
 set_8311_sw_ver() {
@@ -139,11 +155,11 @@ set_8311_sw_ver() {
 }
 
 get_8311_hw_ver() {
-	fw_printenv -n "8311_hw_ver" 2>/dev/null | head -c 14
+	fwenv_get_8311 "hw_ver" | head -c 14
 }
 
 get_8311_cp_hw_ver_sync() {
-	fw_printenv -n 8311_cp_hw_ver_sync 2>/dev/null
+	fwenv_get_8311 "cp_hw_ver_sync" 2>/dev/null
 }
 
 set_8311_hw_ver() {
@@ -164,7 +180,7 @@ set_8311_cp_hw_ver_sync() {
 }
 
 get_8311_equipment_id() {
-	fw_printenv -n 8311_equipment_id 2>/dev/null | head -c 20
+	fwenv_get_8311 "equipment_id" | head -c 20
 }
 
 set_8311_equipment_id() {
@@ -176,7 +192,7 @@ set_8311_equipment_id() {
 
 get_8311_lct_mac() {
 	if [ ! -f "/tmp/8311-lct-mac" ]; then
-		local lct_mac=$(fw_printenv -n 8311_lct_mac 2>/dev/null | grep -i -E '^[0-9a-f]{2}(:[0-9a-f]{2}){5}$')
+		local lct_mac=$(fwenv_get_8311 "lct_mac" | grep -i -E '^[0-9a-f]{2}(:[0-9a-f]{2}){5}$')
 		echo "${lct_mac:-$(pon_mac_get lct)}" | strtoupper > "/tmp/8311-lct-mac"
 	fi
 
@@ -190,7 +206,7 @@ set_8311_lct_mac() {
 
 get_8311_iphost_mac() {
 	if [ ! -f "/tmp/8311-iphost-mac" ]; then
-		local iphost_mac=$(fw_printenv -n 8311_iphost_mac 2>/dev/null | grep -i -E '^[0-9a-f]{2}(:[0-9a-f]{2}){5}$')
+		local iphost_mac=$(fwenv_get_8311 "iphost_mac" 2>/dev/null | grep -i -E '^[0-9a-f]{2}(:[0-9a-f]{2}){5}$')
 		echo "${iphost_mac:-$(pon_mac_get host)}" | strtoupper > "/tmp/8311-iphost-mac"
 	fi
 
@@ -203,7 +219,7 @@ set_8311_iphost_mac() {
 }
 
 get_8311_iphost_hostname() {
-	fw_printenv -n 8311_iphost_hostname 2>/dev/null | head -c 25
+	fwenv_get_8311 "iphost_hostname" | head -c 25
 }
 
 set_8311_iphost_hostname() {
@@ -212,7 +228,7 @@ set_8311_iphost_hostname() {
 }
 
 get_8311_iphost_domain() {
-	fw_printenv -n 8311_iphost_domain 2>/dev/null | head -c 25
+	fwenv_get_8311 "iphost_domain" | head -c 25
 }
 
 set_8311_iphost_domain() {
@@ -221,25 +237,25 @@ set_8311_iphost_domain() {
 }
 
 get_8311_ipaddr() {
-	[ -f "/tmp/8311-ipaddr" ] || { fw_printenv -n 8311_ipaddr 2>/dev/null || echo "192.168.11.1"; } > "/tmp/8311-ipaddr"
+	[ -f "/tmp/8311-ipaddr" ] || { fwenv_get_8311 "ipaddr" || echo "192.168.11.1"; } > "/tmp/8311-ipaddr"
 
 	cat "/tmp/8311-ipaddr"
 }
 
 get_8311_netmask() {
-	[ -f "/tmp/8311-netmask" ] || { fw_printenv -n 8311_netmask 2>/dev/null || echo "255.255.255.0"; } > "/tmp/8311-netmask"
+	[ -f "/tmp/8311-netmask" ] || { fwenv_get_8311 "netmask" || echo "255.255.255.0"; } > "/tmp/8311-netmask"
 
 	cat "/tmp/8311-netmask"
 }
 
 get_8311_gateway() {
-	[ -f "/tmp/8311-gateway" ] || { fw_printenv -n 8311_gateway 2>/dev/null || get_8311_ipaddr; } > "/tmp/8311-gateway"
+	[ -f "/tmp/8311-gateway" ] || { fwenv_get_8311 "gateway" || get_8311_ipaddr; } > "/tmp/8311-gateway"
 
 	cat "/tmp/8311-gateway"
 }
 
 get_8311_loid() {
-	fw_printenv -n 8311_loid 2>/dev/null | head -c 24
+	fwenv_get_8311 "loid" | head -c 24
 }
 
 set_8311_loid() {
@@ -249,7 +265,7 @@ set_8311_loid() {
 }
 
 get_8311_lpwd() {
-	fw_printenv -n 8311_lpwd 2>/dev/null | head -c 12
+	fwenv_get_8311 "lpwd" | head -c 12
 }
 
 set_8311_lpwd() {
@@ -259,7 +275,7 @@ set_8311_lpwd() {
 }
 
 get_8311_ethtool() {
-	fw_printenv -n 8311_ethtool_speed 2>/dev/null
+	fwenv_get_8311 "ethtool_speed"
 }
 
 set_8311_ethtool() {
@@ -268,16 +284,17 @@ set_8311_ethtool() {
 }
 
 get_8311_root_pwhash() {
-	fw_printenv -n 8311_root_pwhash 2>/dev/null | sed 's#/#\\/#g'
+	fwenv_get_8311 "root_pwhash"
 }
 
 set_8311_root_pwhash() {
 	echo "Setting root password hash: $1" | to_console
-	sed -r "s/(root:)([^:]*)(:.+)/\1${1}\3/g" -i /etc/shadow
+	HASH=$(echo "$1" | sed 's#/#\\/#g')
+	sed -r "s/(root:)([^:]*)(:.+)/\1${HASH}\3/g" -i /etc/shadow
 }
 
 get_8311_pon_slot() {
-	fw_printenv -n 8311_pon_slot 2>/dev/null | grep -E '^[0-9]+$'
+	fwenv_get_8311 "pon_slot" | grep -E '^[0-9]+$'
 }
 
 set_8311_pon_slot() {
@@ -292,9 +309,18 @@ set_8311_pon_slot() {
 	fi
 }
 
+get_8311_hostname() {
+	fwenv_get_8311 "hostname"
+}
+
+set_8311_hostname() {
+	echo "Setting system hostname to: $1" | to_console
+	echo "$1" > "/proc/sys/kernel/hostname"
+}
+
 get_8311_ping_host() {
 	if [ ! -f "/tmp/8311-ping-host" ]; then
-		local PING_HOST=$(fw_printenv -n 8311_ping_ip 2>/dev/null)
+		local PING_HOST=$(fwenv_get_8311 "ping_ip")
 
 		if [ -z "$PING_HOST" ]; then
 			local ipaddr=$(get_8311_ipaddr)
@@ -338,11 +364,13 @@ get_8311_module_type() {
 }
 
 get_8311_base_mac() {
-	local serial=$(dd if=/sys/class/pon_mbox/pon_mbox0/device/eeprom50 bs=1 skip=68 count=12 2>/dev/null)
-	local suffix=$(echo "$serial" | tail -c 7 | filterhex)
-	[ -z "$suffix" ] && suffix=$(echo -n "$serial" | sha256sum | head -c 6 | strtoupper)
+	if [ ! -f "/tmp/8311-base-mac" ]; then
+		local serial=$(dd if=/sys/class/pon_mbox/pon_mbox0/device/eeprom50 bs=1 skip=68 count=12 2>/dev/null)
+		local suffix=$(echo "$serial" | tail -c 7 | filterhex)
+		[ -z "$suffix" ] && suffix=$(echo -n "$serial" | sha256sum | head -c 6 | strtoupper)
 
-	echo -n "10:B3:6F"
+		{ echo -n "10:B3:6F"; echo "$suffix" | sed -r 's/(..)/:\1/g'; } > "/tmp/8311-base-mac"
+	fi
 
-	echo "$suffix" | sed -r 's/(..)/:\1/g'
+	cat "/tmp/8311-base-mac"
 }
